LoadedWBCollection Class

A convenience class for handling collections of LoadedWB objects.

See also

Public fields

sheets_list

A list of LoadedWB objects

Active bindings

all_paths

The paths of all workbooks in the collection

all_names

The names of all workbooks in the collection

modified

sheets_list

Methods

Public methods

LoadedWBCollection$new()

LoadedWBCollection$names()

LoadedWBCollection$print()

LoadedWBCollection$clone()

Method new()

Create a new LoadedWBCollection object

Usage

LoadedWBCollection$new(sheets_list = list())

Arguments

sheets_list

A list of LoadedWB objects

Returns

A new LoadedWBCollection object

Method names()

Get the names of the sheets in the collection

Usage

LoadedWBCollection$names()

Returns

A character vector of sheet names

Method print()

Print method for LoadedWBCollection

Usage

LoadedWBCollection$print()

Returns

The LoadedWBCollection object invisibly

Method clone()

The objects of this class are cloneable with this method.

Usage

LoadedWBCollection$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.