Export the flextable to an Excel file with formatting

This function exports a flextable object generated by export_appendix_table_excel to an Excel file, forcing specific formatting such as column widths, row heights, and text rotation.

Usage

export_appendix_table_excel(
  ft,
  file_path,
  sheet_name = "table",
  body_cell_width = 4.5,
  species_height = 100,
  lme_width = 4
)

Arguments

ft

A flextable object to be exported

file_path

The file path where the Excel file will be saved

sheet_name

The name of the worksheet in the Excel file (default is "table")

body_cell_width

The width of the body cells (default is 4)

species_height

The height of the species header row (default is 100)

lme_width

The width of the LME column (default is 4)

Examples

if (FALSE) { # \dontrun{
export_appendix_table_excel(ft, "appendix_table.xlsx")
} # }